git-svn-id: http://jrt.choury.com/svn/myos@13 2cdf69c7-79f9-4451-ab89-d57ae126d6d6

+++ include/file.h
@@ -10,12 +10,12 @@
 #define TTY        1
 
 typedef struct{
-    uint8  isused;
-    uint16 dev;
+    uint32  isused;
+    uint32 dev;
     uint32 indexno;
     uint32 offset;
-    uint16 startnode;
-    uint16 curnode;
+    uint32 startnode;
+    uint32 curnode;
     uint32 length;
 }__attribute__ ((packed)) fileindex;
